# pkg/swagger/docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This API provides programmatic access to sending SMS messages on Android devices. Features include sending SMS, checking message status, device management, webhook configuration, and system health checks.",
        "title": "SMS Gateway for Android™ API",
        "contact": {
            "name": "SMSGate Support",
            "email": "support@sms-gate.app"
        },
        "version": "{APP_VERSION}"
    },
    "host": "api.sms-gate.app",
    "paths": {
...
        "/3rdparty/v1/messages": {
            "get": {
                "security": [
                    {
                        "ApiAuth": []
                    }
                ],
                "description": "Retrieves a list of messages with filtering and pagination",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User",
                    "Messages"
                ],
                "summary": "Get messages",
                "parameters": [
                    {
                        "type": "string",
                        "format": "date-time",
                        "description": "Start date in RFC3339 format",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "date-time",
                        "description": "End date in RFC3339 format",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter messages by processing state",
                        "name": "state",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by device ID",
                        "name": "deviceId",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "Pagination limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Pagination offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of messages",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/smsgateway.MessageState"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiAuth": []
                    }
                ],
                "description": "Enqueues a message for sending. If `deviceId` is set, the specified device is used; otherwise a random registered device is chosen.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User",
                    "Messages"
                ],
                "summary": "Enqueue message",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "Skip phone validation",
                        "name": "skipPhoneValidation",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "default": 0,
                        "description": "Filter devices active within the specified number of hours",
                        "name": "deviceActiveWithin",
                        "in": "query"
                    },
                    {
                        "description": "Send message request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/smsgateway.Message"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Message enqueued",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.GetMessageResponse"
                        },
                        "headers": {
                            "Location": {
                                "type": "string",
                                "description": "Get message state URL"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Message with such ID already exists",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/3rdparty/v1/messages/{id}": {
            "get": {
                "security": [
                    {
                        "ApiAuth": []
                    }
                ],
                "description": "Returns message state by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User",
                    "Messages"
                ],
                "summary": "Get message state",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Message ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Message state",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.GetMessageResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/mobile/v1/message": {
            "get": {
                "security": [
                    {
                        "MobileToken": []
                    }
                ],
                "description": "Returns list of pending messages",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Device",
                    "Messages"
                ],
                "summary": "Get messages for sending",
                "parameters": [
                    {
                        "enum": [
                            "lifo",
                            "fifo"
                        ],
                        "type": "string",
                        "default": "lifo",
                        "description": "Message processing order: lifo (default) or fifo",
                        "name": "order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of pending messages",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/smsgateway.MobileMessage"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "MobileToken": []
                    }
                ],
                "description": "Updates message state",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Device",
                    "Messages"
                ],
                "summary": "Update message state",
                "parameters": [
                    {
                        "description": "List of message state updates",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/smsgateway.MobilePatchMessageItem"
                            }
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successfully updated"
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "smsgateway.DataMessage": {
            "type": "object",
            "required": [
                "data",
                "port"
            ],
            "properties": {
                "data": {
                    "description": "Base64-encoded payload",
                    "type": "string",
                    "format": "byte",
                    "maxLength": 65535,
                    "minLength": 4,
                    "example": "SGVsbG8gV29ybGQh"
                },
                "port": {
                    "description": "Destination port",
                    "type": "integer",
                    "maximum": 65535,
                    "minimum": 1,
                    "example": 53739
                }
            }
        },
...
        "smsgateway.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Error code",
                    "type": "integer"
                },
                "data": {
                    "description": "Error context"
                },
                "message": {
                    "description": "Error message",
                    "type": "string",
                    "example": "An error occurred"
                }
            }
        },
        "smsgateway.GetMessageResponse": {
            "type": "object",
            "required": [
                "deviceId",
                "id",
                "recipients",
                "state"
            ],
            "properties": {
                "deviceId": {
                    "description": "Device ID",
                    "type": "string",
                    "maxLength": 21,
                    "example": "PyDmBQZZXYmyxMwED8Fzy"
                },
                "id": {
                    "description": "Message ID",
                    "type": "string",
                    "maxLength": 36,
                    "example": "PyDmBQZZXYmyxMwED8Fzy"
                },
                "isEncrypted": {
                    "description": "Encrypted",
                    "type": "boolean",
                    "example": false
                },
                "isHashed": {
                    "description": "Hashed",
                    "type": "boolean",
                    "example": false
                },
                "recipients": {
                    "description": "Recipients states",
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/smsgateway.RecipientState"
                    }
                },
                "state": {
                    "description": "State",
                    "allOf": [
                        {
                            "$ref": "#/definitions/smsgateway.ProcessingState"
                        }
                    ],
                    "example": "Pending"
                },
                "states": {
                    "description": "History of states",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
...
        "smsgateway.Message": {
            "type": "object",
            "required": [
                "phoneNumbers"
            ],
            "properties": {
                "dataMessage": {
                    "description": "Data message",
                    "allOf": [
                        {
                            "$ref": "#/definitions/smsgateway.DataMessage"
                        }
                    ]
                },
                "deviceId": {
                    "description": "Optional device ID for explicit selection",
                    "type": "string",
                    "maxLength": 21,
                    "example": "PyDmBQZZXYmyxMwED8Fzy"
                },
                "id": {
                    "description": "ID (if not set - will be generated)",
                    "type": "string",
                    "maxLength": 36,
                    "example": "PyDmBQZZXYmyxMwED8Fzy"
                },
                "isEncrypted": {
                    "description": "Is encrypted",
                    "type": "boolean",
                    "example": true
                },
                "message": {
                    "description": "Message content\nDeprecated: use TextMessage instead",
                    "type": "string",
                    "maxLength": 65535,
                    "example": "Hello World!"
                },
                "phoneNumbers": {
                    "description": "Recipients (phone numbers)",
                    "type": "array",
                    "maxItems": 100,
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "79990001234"
                    ]
                },
                "priority": {
                    "description": "Priority, messages with values greater than `99` will bypass limits and delays",
                    "default": 0,
                    "maximum": 127,
                    "minimum": -128,
                    "allOf": [
                        {
                            "$ref": "#/definitions/smsgateway.MessagePriority"
                        }
                    ],
                    "example": 0
                },
                "simNumber": {
                    "description": "SIM card number (1-3), if not set - default SIM will be used",
                    "type": "integer",
                    "maximum": 3,
                    "example": 1
                },
                "textMessage": {
                    "description": "Text message",
                    "allOf": [
                        {
                            "$ref": "#/definitions/smsgateway.TextMessage"
                        }
                    ]
                },
                "ttl": {
                    "description": "Time to live in seconds (conflicts with `validUntil`)",
                    "type": "integer",
                    "minimum": 5,
                    "example": 86400
                },
                "validUntil": {
                    "description": "Valid until (conflicts with `ttl`)",
                    "type": "string",
                    "example": "2020-01-01T00:00:00Z"
                },
                "withDeliveryReport": {
                    "description": "With delivery report",
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "smsgateway.MessagePriority": {
            "type": "integer",
            "enum": [
                -128,
                0,
                100,
                127
            ],
            "x-enum-comments": {
                "PriorityBypassThreshold": "Threshold at which messages bypass limits and delays"
            },
            "x-enum-varnames": [
                "PriorityMinimum",
                "PriorityDefault",
                "PriorityBypassThreshold",
                "PriorityMaximum"
            ]
        },
        "smsgateway.MessageState": {
            "type": "object",
            "required": [
                "deviceId",
                "id",
                "recipients",
                "state"
            ],
            "properties": {
                "deviceId": {
                    "description": "Device ID",
                    "type": "string",
                    "maxLength": 21,
                    "example": "PyDmBQZZXYmyxMwED8Fzy"
                },
                "id": {
                    "description": "Message ID",
                    "type": "string",
                    "maxLength": 36,
                    "example": "PyDmBQZZXYmyxMwED8Fzy"
                },
                "isEncrypted": {
                    "description": "Encrypted",
                    "type": "boolean",
                    "example": false
                },
                "isHashed": {
                    "description": "Hashed",
                    "type": "boolean",
                    "example": false
                },
                "recipients": {
                    "description": "Recipients states",
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/smsgateway.RecipientState"
                    }
                },
                "state": {
                    "description": "State",
                    "allOf": [
                        {
                            "$ref": "#/definitions/smsgateway.ProcessingState"
                        }
                    ],
                    "example": "Pending"
                },
                "states": {
                    "description": "History of states",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
...
        "smsgateway.MobileMessage": {
            "type": "object",
            "required": [
                "phoneNumbers"
            ],
            "properties": {
                "createdAt": {
                    "description": "Message creation time",
                    "type": "string",
                    "example": "2020-01-01T00:00:00Z"
                },
                "dataMessage": {
                    "description": "Data message",
                    "allOf": [
                        {
                            "$ref": "#/definitions/smsgateway.DataMessage"
                        }
                    ]
                },
                "deviceId": {
                    "description": "Optional device ID for explicit selection",
                    "type": "string",
                    "maxLength": 21,
                    "example": "PyDmBQZZXYmyxMwED8Fzy"
                },
                "id": {
                    "description": "ID (if not set - will be generated)",
                    "type": "string",
                    "maxLength": 36,
                    "example": "PyDmBQZZXYmyxMwED8Fzy"
                },
                "isEncrypted": {
                    "description": "Is encrypted",
                    "type": "boolean",
                    "example": true
                },
                "message": {
                    "description": "Message content\nDeprecated: use TextMessage instead",
                    "type": "string",
                    "maxLength": 65535,
                    "example": "Hello World!"
                },
                "phoneNumbers": {
                    "description": "Recipients (phone numbers)",
                    "type": "array",
                    "maxItems": 100,
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "79990001234"
                    ]
                },
                "priority": {
                    "description": "Priority, messages with values greater than `99` will bypass limits and delays",
                    "default": 0,
                    "maximum": 127,
                    "minimum": -128,
                    "allOf": [
                        {
                            "$ref": "#/definitions/smsgateway.MessagePriority"
                        }
                    ],
                    "example": 0
                },
                "simNumber": {
                    "description": "SIM card number (1-3), if not set - default SIM will be used",
                    "type": "integer",
                    "maximum": 3,
                    "example": 1
                },
                "textMessage": {
                    "description": "Text message",
                    "allOf": [
                        {
                            "$ref": "#/definitions/smsgateway.TextMessage"
                        }
                    ]
                },
                "ttl": {
                    "description": "Time to live in seconds (conflicts with `validUntil`)",
                    "type": "integer",
                    "minimum": 5,
                    "example": 86400
                },
                "validUntil": {
                    "description": "Valid until (conflicts with `ttl`)",
                    "type": "string",
                    "example": "2020-01-01T00:00:00Z"
                },
                "withDeliveryReport": {
                    "description": "With delivery report",
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "smsgateway.MobilePatchMessageItem": {
            "type": "object",
            "required": [
                "id",
                "recipients",
                "state"
            ],
            "properties": {
                "id": {
                    "description": "Message ID",
                    "type": "string",
                    "maxLength": 36,
                    "example": "PyDmBQZZXYmyxMwED8Fzy"
                },
                "recipients": {
                    "description": "Recipients states",
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/smsgateway.RecipientState"
                    }
                },
                "state": {
                    "description": "State",
                    "allOf": [
                        {
                            "$ref": "#/definitions/smsgateway.ProcessingState"
                        }
                    ],
                    "example": "Pending"
                },
                "states": {
                    "description": "History of states",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
...
        "smsgateway.ProcessingState": {
            "type": "string",
            "enum": [
                "Pending",
                "Processed",
                "Sent",
                "Delivered",
                "Failed"
            ],
            "x-enum-comments": {
                "ProcessingStateDelivered": "Delivered",
                "ProcessingStateFailed": "Failed",
                "ProcessingStatePending": "Pending",
                "ProcessingStateProcessed": "Processed (received by device)",
                "ProcessingStateSent": "Sent"
            },
            "x-enum-varnames": [
                "ProcessingStatePending",
                "ProcessingStateProcessed",
                "ProcessingStateSent",
                "ProcessingStateDelivered",
                "ProcessingStateFailed"
            ]
        },
...
        "smsgateway.RecipientState": {
            "type": "object",
            "required": [
                "phoneNumber",
                "state"
            ],
            "properties": {
                "error": {
                    "description": "Error (for `Failed` state)",
                    "type": "string",
                    "example": "timeout"
                },
                "phoneNumber": {
                    "description": "Phone number or first 16 symbols of SHA256 hash",
                    "type": "string",
                    "maxLength": 128,
                    "minLength": 1,
                    "example": "79990001234"
                },
                "state": {
                    "description": "State",
                    "allOf": [
                        {
                            "$ref": "#/definitions/smsgateway.ProcessingState"
                        }
                    ],
                    "example": "Pending"
                }
            }
        },
...
        "smsgateway.TextMessage": {
            "type": "object",
            "required": [
                "text"
            ],
            "properties": {
                "text": {
                    "description": "Message text",
                    "type": "string",
                    "maxLength": 65535,
                    "minLength": 1,
                    "example": "Hello World!"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiAuth": {
            "type": "basic"
        },
        "MobileToken": {
            "description": "Mobile device token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "ServerKey": {
            "description": "Private server authentication",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "UserCode": {
            "description": "User one-time code authentication",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}